Feat: Movie 모듈의 Movie 엔티티를 삭제할때 조건 추가 및 통합 이벤트 발행

+++ Movie.API/Application/Commands/DeleteMovieCommandHandler.cs
@@ -1,3 +1,5 @@
+using IntegrationEvents;
+using IntegrationEvents.Events;
 using MediatR;
 using Movie.Domain.Aggregate;
 using Movie.Domain.Exceptions;
@@ -5,15 +7,28 @@ using movie_shop_asp.Server.Movie.API.Application.Commands;
 
 namespace Movie.API.Application.Commands;
 
-public class DeleteMovieCommandHandler(IMovieRepository movieRepository) : IRequestHandler<DeleteMovieCommand, bool>
+public class DeleteMovieCommandHandler(IMovieRepository movieRepository, InProcessIntegrationEventService integrationEventService) : IRequestHandler<DeleteMovieCommand, bool>
 {
     public async Task<bool> Handle(DeleteMovieCommand request, CancellationToken cancellationToken)
     {
         var movie = await movieRepository.GetAsync(request.MovieId) 
             ?? throw new MovieDomainException("대상 영화를 찾을 수 없습니다.");
 
-        movieRepository.Delete(movie);
+        if (!movie.CanRemove())
+        {
+            throw new MovieDomainException("상영중에는 영화를 삭제할 수 없습니다.");
+        }
 
-        return await movieRepository.UnitOfWork.SaveEntitiesAsync(cancellationToken);
+        movieRepository.Delete(movie);
+        await movieRepository.UnitOfWork.SaveEntitiesAsync(cancellationToken);
+
+        var integrationEvent = new MovieDeletedIntegrationEvent
+        {
+            MovieId = request.MovieId
+        };
+
+        integrationEventService.Add(integrationEvent);
+
+        return true;
     }
 }
